spearate concepts into detail
